feature: criado opcao de criar agendamento e recarregar horas disponiveis

+++ src/module/schedule/cancel-schedule.js
@@ -1,20 +1,23 @@
 import {deleteSchedule} from "../../services/delete-schedule.js"
 import {loadSchedules} from "./load-schedules.js"
+const sectionForm = document.getElementById("new-schedule");
 
 export async function cancelSchedule(target, id) {
     target.addEventListener("click", async () => {
-        try{
-            const isConfirm = confirm("Deseja realmente deletar esse agendamento?");
-
-            if(isConfirm){
-                await deleteSchedule(id);
-                loadSchedules();
-                alert("Agendamento deletado com sucesso!");
+        if (sectionForm.classList.contains("display-none")){
+            try{
+                const isConfirm = confirm("Deseja realmente deletar esse agendamento?");
+    
+                if(isConfirm){
+                    await deleteSchedule(id);
+                    loadSchedules();
+                    alert("Agendamento deletado com sucesso!");
+                }
             }
-        }
-        catch(error) {
-            alert("Erro ao deletar Agendamento!");
-            console.log(error);
+            catch(error) {
+                alert("Erro ao deletar Agendamento!");
+                console.log(error);
+            }
         }
     })
 }
